refactor(api): decouple lazy loading from request factory internals
Prepares AppSyncGraphQLRequestFactory to move to aws-api-appsync (Phase 1) by
removing the module-internal coupling between the factory and the plugin's
lazy-loading types. No behavior change and no public API change.

- Add buildQueryFromKeyMap, absorbing the schema-to-variables loop that was
  duplicated in ApiLazyModelReference. GraphQLRequestVariable is no longer
  referenced outside the factory, so it can move with it while staying internal.
- buildQueryInternal is now private; nothing outside the factory called it once
  the lazy path stopped.
- buildModelPageQuery takes the page type from the caller rather than
  referencing ApiModelPage, so the factory no longer depends on any
  lazy-loading type. This is what removes the dependency cycle that would
  otherwise block the move.

Both new entry points are @InternalAmplifyApi and so are excluded from the
binary compatibility dump; apiCheck passes with no apiDump.
